updated activity card
